+++ middleware/query.js
@@ -5,13 +5,13 @@ var middleware = module.exports = {
     head: function (request, response, next) {
       var Model = request.app.get('model');
       request.baucis.noBody = true;
-      request.baucis.query = Model.findOne(request.app.getFindByCondition(request));
+      request.baucis.query = Model.findOne(request.app.getFindByConditions(request));
       next();
     },
     // Retrieve the addressed document
     get: function (request, response, next) {
       var Model = request.app.get('model');
-      request.baucis.query = Model.findOne(request.app.getFindByCondition(request));
+      request.baucis.query = Model.findOne(request.app.getFindByConditions(request));
       next();
     },
     // Treat the addressed document as a collection, and push
@@ -27,13 +27,13 @@ var middleware = module.exports = {
       if (bodyId && request.params.id !== bodyId) return next(new Error('ID mismatch'));
 
       request.baucis.updateWithBody = true;
-      request.baucis.query = Model.findOne(request.app.getFindByCondition(request));
+      request.baucis.query = Model.findOne(request.app.getFindByConditions(request));
       next();
     },
     // Delete the addressed object
     del: function (request, response, next) {
       var Model = request.app.get('model');
-      request.baucis.query = Model.remove(request.app.getFindByCondition(request));
+      request.baucis.query = Model.remove(request.app.getFindByConditions(request));
       next();
     }
   },